refactor: split MarkdownTaskClient into focused classes for better testability
Split the 960-line MarkdownTaskClient into 4 classes with single responsibilities:
- MarkdownParser (408 lines): Parse Markdown to extract tasks
- MarkdownSerializer (413 lines): Apply edits to Markdown (create/update/delete/move)
- TaskLineEditor (121 lines): Edit task lines (title/status/metadata)
- MarkdownTaskClient (54 lines): Facade maintaining backward compatibility

This improves testability by:
- Reducing cyclomatic complexity per file
- Enabling isolated unit testing of each responsibility
- Making mock dependencies clearer
