+++ packages/core/src/infrastructure/clients/index.ts
@@ -10,6 +10,11 @@ export {
 	SerializerError,
 	type TaskEdit,
 } from './markdownTaskClient';
+
+// Separated classes for fine-grained testing
+export { MarkdownParser } from './markdownParser';
+export { MarkdownSerializer } from './markdownSerializer';
+export { TaskLineEditor } from './taskLineEditor';
 export { type FrontmatterResult, RemarkClient } from './remarkClient';
 export { VscodeConfigClient, type VscodeConfigDeps } from './vscodeConfigClient';
 export {
